update vm_int_t to 32bit as LP64

+++ include/processwarp/processwarp.h
@@ -10,9 +10,9 @@ extern "C" {
   typedef uint64_t __pw_vm_ptr_t;
 
   /** int with VM */
-  typedef int64_t __pw_vm_int_t;
+  typedef int32_t __pw_vm_int_t;
   /** unsigned int with VM */
-  typedef uint64_t __pw_vm_uint_t;
+  typedef uint32_t __pw_vm_uint_t;
 
   __pw_vm_int_t at_after_warp(void (*func)(void));
